fixing inc comparison btw dif types

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -9,7 +9,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *new_table = NULL;
 	hash_node_t **array = NULL;
-	unsigned int iter;
+	unsigned long int iter;
 
 	new_table = malloc(sizeof(hash_table_t));
 	if (new_table == NULL)
